Add VHDL_SOURCES_lib (GHDL only)
Currently, to include external libraries, the user has to manually run
ghdl before launching cocotb. With this commit, it is now possible to
specify in which library a set of VHDL files has to be placed into.
To do so, the files can be specified directly in the makefile under
the VHDL_SOURCES_lib variable. For example:

VHDL_SOURCES = top.vhd
VHDL_SOURCES_mylib = libfile.vhd libfile2.vhd

Here, libfile.vhd and libfile2.vhd will be analyzed with the
--work=mylib flag in GHDL before analyzing top.vhd.